Moved all StoreManagement related Kotlin files to StoreManagement package.
