+++ android-app/app/src/main/java/com/example/fieldsync/MainActivity.kt
@@ -6,6 +6,7 @@ import androidx.appcompat.app.AppCompatActivity
 
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.fieldsync.StoreManagement.StoreManagement
 import com.example.fieldsync.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
