+++ android-app/app/src/main/java/com/example/fieldsync/StoreManagement/VisitLocation.kt
@@ -1,4 +1,4 @@
-package com.example.fieldsync
+package com.example.fieldsync.StoreManagement
 
 data class VisitLocation(
     val name: String,
