+++ android-app/app/src/main/res/layout/activity_store_management.xml
@@ -6,7 +6,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".StoreManagement">
+        tools:context=".StoreManagement.StoreManagement">
 
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="409dp"
